- Cleaned up initializers - Added factories and other convenience methods - Attempted to deal with synchronization. If a specific #define is 1 (will now use pthread_mutex_lock() instead of @synchronized as it is faster by orders of measure - see http://googlemac.blogspot.com/2006/10/synchronized-swimming.html for details), - Implemented preliminary template system (still need to implement the dictionary based saturation method). - A lot of time went into hardening and bug fixing the accessor methods as those are crucial (still need to test for leaks with ObjectAlloc and/or Instruments).
Currently 10.4 is supported, but move to 10.5 may be justified (at the moment this is driven by the build requirements for Quicksilver, seeing as it can be switched to use 10.5 SDK and will build OK we may finally be able to use 10.5 as base for BMScript too). 

Also went from using BMScript as abstract superclass to a normal instantiate-able class since the ceonvenience/factories make it flexible in its usage to support a multitude of scripting languages. 
Nevertheless BMScriptLanguageProtocol stays in place to ease subclassing should the need arise. 

The behaviour for -alloc init is now defined by the implemented required methods of the protocol since BMScript now conforms to that protocol. 
Delegation has been built in, as well as background spawning tasks (BMScript's state keeps track of both the blocking task and the non-blocking task, in the hopes of avoiding shared state and minimizing threading issues)
For this notifications have been both used and build into BMScript.

TODO:

- Get more test coverage. 
- Implement dictionary based template method/system
- Understand NSAutoreleasePools better