Move the check for scriptOptions so that it happens before checking the scriptSource.
This allows the delegate callbacks employed throughout the scriptOptions check to saturate
the options ivar. If the options ivars is filled this will further keep BMScript from
assuming it has to supply an arbitrary default scriptSource and just assure that the normal
machinery is used (e.g. check for scriptSource -> ask delegate in callbacks -> assume nil and
supply default source since we probably also do not have any options set we launch with /bin/echo).