Change BMScript to return task results exclusively as NSData.
Also update examples, tests, unit tests and documentation.

Ratio

Up to now I went on assuming that UTF-8 encoding the data returned
by the underlying task might work about 80% of the time. But I knew 
eventually I had to redo it properly. 

While planned for the v0.2 release of BMScript it just didn't make it 
quite in there. I wanted to play around with a few more ideas such as 
making the results dynamically typed or keeping results as string but 
making partialResult as data and public so the caller can peek inside 
it if the string result is nil. 
I know all of these where wrong but I considered it only in a "just works" 
spirit, but as NSString aptly demonstrates - it is just not possible. 

So I went on with breaking existing tests and usages - in the end I 
decided the only proper way to do it is to return the data unmodified 
since the task results could be anything. I then added a category method 
to NSData called contentsAsString which will try to convert to a string 
using UTF-8 encoding and if it fails it will give you the description 
string of the data instead. This was to bridge the gap between new and 
old behavior and it is strictly intended for logging purposes. 
For turning critical result data into strings you should perform the 
conversion yourself using the proper encoding.